# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Trueke Bogot\u00E1 Solidario API",
    "description": "Errores: application/problem\u002Bjson (title = mensaje mostrable, errors = validaci\u00F3n por campo). Fechas en UTC ISO-8601 con Z. Enums como texto. Notificaciones en tiempo real: SignalR en /hubs/notificaciones (token por ?access_token=, evento \u0022notificacion\u0022).",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/admin/usuarios/{usuarioId}/rol": {
      "patch": {
        "tags": [
          "Administracion"
        ],
        "parameters": [
          {
            "name": "usuarioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CambiarRolRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CambiarRolRequest"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/CambiarRolRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/denuncias": {
      "post": {
        "tags": [
          "Denuncias"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearDenunciaRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearDenunciaRequest"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/CrearDenunciaRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DenunciaCreadaDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DenunciaCreadaDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DenunciaCreadaDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/denuncias": {
      "get": {
        "tags": [
          "Denuncias"
        ],
        "parameters": [
          {
            "name": "Estado",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/EstadoDenunciaDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DenunciaAgrupadaDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DenunciaAgrupadaDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DenunciaAgrupadaDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/denuncias/{id}/resolver": {
      "post": {
        "tags": [
          "Denuncias"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolverDenunciaRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolverDenunciaRequest"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/ResolverDenunciaRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/pagos/iniciar": {
      "post": {
        "tags": [
          "Pagos"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IniciarPagoRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/IniciarPagoRequest"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/IniciarPagoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagoIniciadoDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagoIniciadoDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagoIniciadoDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pagos/{referencia}": {
      "get": {
        "tags": [
          "Pagos"
        ],
        "parameters": [
          {
            "name": "referencia",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagoEstadoDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagoEstadoDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagoEstadoDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/pagos/{referencia}/simular": {
      "post": {
        "tags": [
          "Pagos"
        ],
        "parameters": [
          {
            "name": "referencia",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "aprobado",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagoEstadoDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagoEstadoDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagoEstadoDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/usuarios/{id}/perfil": {
      "get": {
        "tags": [
          "Perfiles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PerfilUsuarioDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerfilUsuarioDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerfilUsuarioDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/usuarios/{id}/publicaciones": {
      "get": {
        "tags": [
          "Perfiles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Pagina",
            "in": "query",
            "schema": {
              "maximum": 10000,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Tamano",
            "in": "query",
            "schema": {
              "maximum": 50,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionDtoPaginaDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionDtoPaginaDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionDtoPaginaDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/usuarios/{id}/calificaciones": {
      "get": {
        "tags": [
          "Perfiles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Pagina",
            "in": "query",
            "schema": {
              "maximum": 10000,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Tamano",
            "in": "query",
            "schema": {
              "maximum": 50,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CalificacionDtoPaginaDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalificacionDtoPaginaDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalificacionDtoPaginaDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/publicaciones": {
      "get": {
        "tags": [
          "Publicaciones"
        ],
        "parameters": [
          {
            "name": "CategoriaId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Modo",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ModoDto"
            }
          },
          {
            "name": "Localidad",
            "in": "query",
            "schema": {
              "maxLength": 60,
              "type": "string"
            }
          },
          {
            "name": "Texto",
            "in": "query",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "PrecioMin",
            "in": "query",
            "schema": {
              "maximum": 1000000000,
              "minimum": 0,
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "PrecioMax",
            "in": "query",
            "schema": {
              "maximum": 1000000000,
              "minimum": 0,
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "SoloVerificados",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Orden",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/OrdenPublicacionesDto"
            }
          },
          {
            "name": "Pagina",
            "in": "query",
            "schema": {
              "maximum": 10000,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Tamano",
            "in": "query",
            "schema": {
              "maximum": 50,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionDtoPaginaDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionDtoPaginaDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionDtoPaginaDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Publicaciones"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearPublicacionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearPublicacionRequest"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/CrearPublicacionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccionDenunciaDto": {
        "enum": [
          "Descartar",
          "OcultarContenido",
          "MarcarRevisada"
        ],
        "type": "string"
      },
...
      "CalificacionDtoPaginaDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CalificacionDto"
            }
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "pagina": {
            "type": "integer",
            "format": "int32"
          },
          "tamano": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "CambiarRolRequest": {
        "type": "object",
        "properties": {
          "rol": {
            "$ref": "#/components/schemas/RolDto"
          }
        },
        "additionalProperties": false
      },
...
      "ConceptoPagoDto": {
        "enum": [
          "Destacar",
          "Verificar",
          "Premium",
          "Empresa",
          "Recarga"
        ],
        "type": "string"
      },
...
      "CotizacionDto": {
        "type": "object",
        "properties": {
          "precioBaseCop": {
            "type": "integer",
            "format": "int32"
          },
          "descuentoPorcentaje": {
            "type": "integer",
            "format": "int32"
          },
          "puntosACanjear": {
            "type": "integer",
            "format": "int32"
          },
          "totalCop": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "CrearDenunciaRequest": {
        "type": "object",
        "properties": {
          "tipo": {
            "$ref": "#/components/schemas/TipoDenunciaDto"
          },
          "objetivoId": {
            "type": "string",
            "format": "uuid"
          },
          "motivo": {
            "$ref": "#/components/schemas/MotivoDenunciaDto"
          },
          "detalle": {
            "maxLength": 500,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CrearPublicacionRequest": {
        "required": [
          "descripcion",
          "localidad",
          "titulo"
        ],
        "type": "object",
        "properties": {
          "titulo": {
            "maxLength": 120,
            "minLength": 3,
            "type": "string"
          },
          "descripcion": {
            "maxLength": 2000,
            "minLength": 1,
            "type": "string"
          },
          "categoriaId": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "modo": {
            "$ref": "#/components/schemas/ModoDto"
          },
          "localidad": {
            "maxLength": 60,
            "minLength": 2,
            "type": "string"
          },
          "precioReferenciaCop": {
            "maximum": 1000000000,
            "minimum": 0.01,
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "latitud": {
            "maximum": 90,
            "minimum": -90,
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "longitud": {
            "maximum": 180,
            "minimum": -180,
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "imagenes": {
            "maxItems": 5,
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DenunciaAgrupadaDto": {
        "type": "object",
        "properties": {
          "denunciaId": {
            "type": "string",
            "format": "uuid"
          },
          "tipo": {
            "type": "string"
          },
          "objetivoId": {
            "type": "string",
            "format": "uuid"
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "motivos": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "detalles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "vistaPrevia": {
            "type": "string",
            "nullable": true
          },
          "objetivoExiste": {
            "type": "boolean"
          },
          "primeraUtc": {
            "type": "string",
            "format": "date-time"
          },
          "ultimaUtc": {
            "type": "string",
            "format": "date-time"
          },
          "estado": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "DenunciaCreadaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "estado": {
            "type": "string"
          },
          "fechaUtc": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "EstadoDenunciaDto": {
        "enum": [
          "Pendiente",
          "Resuelta",
          "Descartada"
        ],
        "type": "string"
      },
...
      "IniciarPagoRequest": {
        "type": "object",
        "properties": {
          "concepto": {
            "$ref": "#/components/schemas/ConceptoPagoDto"
          },
          "publicacionId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "documentoUrl": {
            "maxLength": 500,
            "type": "string",
            "nullable": true
          },
          "montoRecargaCop": {
            "maximum": 10000000,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ModoDto": {
        "enum": [
          "Trueke",
          "Compra",
          "Donacion"
        ],
        "type": "string"
      },
      "MotivoDenunciaDto": {
        "enum": [
          "Spam",
          "Fraude",
          "ContenidoInapropiado",
          "ArticuloProhibido",
          "Acoso",
          "Otro"
        ],
        "type": "string"
      },
...
      "OrdenPublicacionesDto": {
        "enum": [
          "Recientes",
          "PrecioAsc",
          "PrecioDesc"
        ],
        "type": "string"
      },
...
      "PagoEstadoDto": {
        "type": "object",
        "properties": {
          "referencia": {
            "type": "string"
          },
          "concepto": {
            "type": "string"
          },
          "montoCop": {
            "type": "integer",
            "format": "int32"
          },
          "estado": {
            "type": "string"
          },
          "fechaUtc": {
            "type": "string",
            "format": "date-time"
          },
          "fechaResolucionUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PagoIniciadoDto": {
        "type": "object",
        "properties": {
          "referencia": {
            "type": "string"
          },
          "concepto": {
            "type": "string"
          },
          "montoCop": {
            "type": "integer",
            "format": "int32"
          },
          "montoEnCentavos": {
            "type": "integer",
            "format": "int64"
          },
          "moneda": {
            "type": "string"
          },
          "proveedor": {
            "type": "string"
          },
          "llavePublica": {
            "type": "string",
            "nullable": true
          },
          "firmaIntegridad": {
            "type": "string",
            "nullable": true
          },
          "cotizacion": {
            "$ref": "#/components/schemas/CotizacionDto"
          },
          "expiraUtc": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "PerfilUsuarioDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nombre": {
            "type": "string"
          },
          "localidad": {
            "type": "string"
          },
          "reputacion": {
            "type": "number",
            "format": "double"
          },
          "verificado": {
            "type": "boolean"
          },
          "tipoCuenta": {
            "type": "string"
          },
          "miembroDesde": {
            "type": "string",
            "format": "date-time"
          },
          "truekesCompletados": {
            "type": "integer",
            "format": "int32"
          },
          "comprasRealizadas": {
            "type": "integer",
            "format": "int32"
          },
          "donacionesRealizadas": {
            "type": "integer",
            "format": "int32"
          },
          "calificacionPromedio": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "totalCalificaciones": {
            "type": "integer",
            "format": "int32"
          },
          "publicacionesActivas": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "PublicacionDtoPaginaDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicacionDto"
            }
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "pagina": {
            "type": "integer",
            "format": "int32"
          },
          "tamano": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ResolverDenunciaRequest": {
        "type": "object",
        "properties": {
          "accion": {
            "$ref": "#/components/schemas/AccionDenunciaDto"
          },
          "nota": {
            "maxLength": 300,
            "minLength": 3,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RolDto": {
        "enum": [
          "Cliente",
          "Administrador",
          "SuperUsuario"
        ],
        "type": "string"
      },
...
      "TipoDenunciaDto": {
        "enum": [
          "Publicacion",
          "Comentario",
          "Mensaje",
          "Usuario",
          "Calificacion"
        ],
        "type": "string"
      },
...
      "UsuarioDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nombreCompleto": {
            "type": "string"
          },
          "localidad": {
            "type": "string"
          },
          "correo": {
            "type": "string"
          },
          "rol": {
            "type": "string"
          },
          "tipoCuenta": {
            "type": "string"
          },
          "planVigenteHasta": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "verificado": {
            "type": "boolean"
          },
          "estadoVerificacion": {
            "type": "string"
          },
          "saldoEcoPuntos": {
            "type": "integer",
            "format": "int32"
          },
          "reputacion": {
            "type": "number",
            "format": "double"
          },
          "truekesCompletados": {
            "type": "integer",
            "format": "int32"
          },
          "comprasRealizadas": {
            "type": "integer",
            "format": "int32"
          },
          "donacionesRealizadas": {
            "type": "integer",
            "format": "int32"
          },
          "destacadosGratisRestantes": {
            "type": "integer",
            "format": "int32"
          },
          "correoVerificado": {
            "type": "boolean"
          },
          "tieneClave": {
            "type": "boolean"
          },
          "vinculadoGoogle": {
            "type": "boolean"
          },
          "dosFactoresActivo": {
            "type": "boolean"
          },
          "codigosRecuperacionRestantes": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Pega el token devuelto por /auth/login (sin la palabra Bearer).",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}